Add category links
