+++ lib/api.js
@@ -1,4 +1,3 @@
-/* global fetch */
 import { airtableGet } from './airtable'
 
 export function fetchAllocations() {
@@ -19,7 +18,7 @@ export function fetchAwards() {
 
 export function fetchCategories() {
   return airtableGet('Categories', {
-    'fields[]': ['Name', 'Description', 'Parent Category', 'Ballot Allocation'],
+    'fields[]': ['Name', 'Description', 'Parent Category', 'Ballot Allocation', 'Documents'],
     filterByFormula: 'NOT({Name} = "")',
     'sort[0][field]': 'Name'
   })
