+++ lib/util.js
@@ -451,6 +451,10 @@ export function getGranteeByProject(project, grantees) {
   return grantees.find(g => g.id === project.fields.Grantee[0])
 }
 
+export function getDocumentById(id, documents) {
+  return documents.find(d => d.id === id)
+}
+
 export function getProjectById(id, projects) {
   return projects.find(p => p.id === id)
 }
